Show popups on node boxes everywhere
The patch that added popups on node boxes skipped the
progress and live deployment views. This patch makes sure
the popups are displayed on all the node boxes on all views.

It also makes sure that the nodes information is available
even when there is no stack -- for the undeploy progress page.

Change-Id: Ib5bd428f9ba4d7b324c2b0fa4421994c4e3165c7